Fixed condition code behaviour and minor improvements
Am2900 Data Book 2-88, Table 5:
Inverting Am2904's CT to avoid confusion so we can use column "CT = H"
and ensure TUM MI machine compliance. This means that we can test for
'is-zero' for equality and 'not-zero' for inequality.
In other words, we want to avoid the Table 5 bottom note's consequences:
doing it the counterintuitive way.